// src/api/openapi/auto-generated-admin/models/PaginatedRoleResultResponse.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter Admin Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RoleResponse } from './RoleResponse';
import {
    RoleResponseFromJSON,
    RoleResponseFromJSONTyped,
    RoleResponseToJSON,
} from './RoleResponse';

/**
 * 
 * @export
 * @interface PaginatedRoleResultResponse
 */
export interface PaginatedRoleResultResponse {
    /**
     * A URL to the JSON Schema for this object.
     * @type {string}
     * @memberof PaginatedRoleResultResponse
     */
    readonly $schema?: string;
    /**
     * 
     * @type {boolean}
     * @memberof PaginatedRoleResultResponse
     */
    hasNextPage: boolean;
    /**
     * 
     * @type {Array<RoleResponse>}
     * @memberof PaginatedRoleResultResponse
     */
    items: Array<RoleResponse>;
    /**
     * 
     * @type {string}
     * @memberof PaginatedRoleResultResponse
     */
    nextCursor: string;
}

/**
 * Check if a given object implements the PaginatedRoleResultResponse interface.
 */
export function instanceOfPaginatedRoleResultResponse(value: object): value is PaginatedRoleResultResponse {
    if (!('hasNextPage' in value) || value['hasNextPage'] === undefined) return false;
    if (!('items' in value) || value['items'] === undefined) return false;
    if (!('nextCursor' in value) || value['nextCursor'] === undefined) return false;
    return true;
}

export function PaginatedRoleResultResponseFromJSON(json: any): PaginatedRoleResultResponse {
    return PaginatedRoleResultResponseFromJSONTyped(json, false);
}

export function PaginatedRoleResultResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): PaginatedRoleResultResponse {
    if (json == null) {
        return json;
    }
    return {
        
        '$schema': json['$schema'] == null ? undefined : json['$schema'],
        'hasNextPage': json['has_next_page'],
        'items': ((json['items'] as Array<any>).map(RoleResponseFromJSON)),
        'nextCursor': json['next_cursor'],
    };
}

export function PaginatedRoleResultResponseToJSON(value?: Omit<PaginatedRoleResultResponse, '$schema'> | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'has_next_page': value['hasNextPage'],
        'items': ((value['items'] as Array<any>).map(RoleResponseToJSON)),
        'next_cursor': value['nextCursor'],
    };
}