// src/api/openapi/auto-generated-admin/models/RoleResponse.ts
/* tslint:disable */
/* eslint-disable */ // @ts-nocheck
/**
 * Jupiter Admin Service
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { Role } from './Role';
import {
    RoleFromJSON,
    RoleFromJSONTyped,
    RoleToJSON,
} from './Role';

/**
 * 
 * @export
 * @interface RoleResponse
 */
export interface RoleResponse {
    /**
     * 
     * @type {Array<string>}
     * @memberof RoleResponse
     */
    members: Array<string>;
    /**
     * 
     * @type {Role}
     * @memberof RoleResponse
     */
    role: Role;
    /**
     * 
     * @type {Array<string>}
     * @memberof RoleResponse
     */
    roleRelations: Array<string>;
}

/**
 * Check if a given object implements the RoleResponse interface.
 */
export function instanceOfRoleResponse(value: object): value is RoleResponse {
    if (!('members' in value) || value['members'] === undefined) return false;
    if (!('role' in value) || value['role'] === undefined) return false;
    if (!('roleRelations' in value) || value['roleRelations'] === undefined) return false;
    return true;
}

export function RoleResponseFromJSON(json: any): RoleResponse {
    return RoleResponseFromJSONTyped(json, false);
}

export function RoleResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): RoleResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'members': json['members'],
        'role': RoleFromJSON(json['role']),
        'roleRelations': json['role_relations'],
    };
}

export function RoleResponseToJSON(value?: RoleResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'members': value['members'],
        'role': RoleToJSON(value['role']),
        'role_relations': value['roleRelations'],
    };
}